- create a new abstract model/service class for the list addition functionality
