+++ lending-admin-client/src/app/add-multiple-games/add-multiple-games.component.html
@@ -1,22 +1,15 @@
 <div class="container">
   <wizard>
-    <lending-file-selection-step wizardStep stepTitle="Datei auswählen"
-                                 (onFileSelection)="loadWorkbook($event)"></lending-file-selection-step>
+    <lending-file-selection-step wizardStep stepTitle="Datei auswählen"></lending-file-selection-step>
 
-    <lending-sheet-selection-step wizardStep stepTitle="Tabelle auswählen" [file]="file"
-                                  (onSheetSelection)="loadSheet($event)"></lending-sheet-selection-step>
+    <lending-sheet-selection-step wizardStep stepTitle="Tabelle auswählen"></lending-sheet-selection-step>
 
-    <lending-check-sheet-content-step wizardStep stepTitle="Tabelleninhalt" [data]="data"
-                                      (onContentChecked)="loadTable($event)"></lending-check-sheet-content-step>
+    <lending-check-sheet-content-step wizardStep stepTitle="Tabelleninhalt"></lending-check-sheet-content-step>
 
-    <lending-column-assignment-step wizardStep stepTitle="Spaltenzuordnung" [fileHeader]="fileHeader" [fileData]="fileContent"
-                                    [allDatabaseHeader]="allDatabaseHeaders" (onColumnAssigned)="loadAssignments($event)"></lending-column-assignment-step>
+    <lending-column-assignment-step wizardStep stepTitle="Spaltenzuordnung"></lending-column-assignment-step>
 
-    <lending-confirmation-step wizardStep stepTitle="Bestätigen" [items]="games" [itemValidator]="verifyGames"
-                               [columns]="columns" [columnHeaders]="columnHeaders"
-                               (onItemsConfirmed)="insertGames($event)"></lending-confirmation-step>
+    <lending-confirmation-step wizardStep stepTitle="Bestätigen"></lending-confirmation-step>
 
-    <lending-summary-step wizardCompletionStep stepTitle="Ergebnis"
-                          [items]="games" [addItemsResult]="insertionResult" (onReset)="reset()"></lending-summary-step>
+    <lending-summary-step wizardCompletionStep stepTitle="Ergebnis"></lending-summary-step>
   </wizard>
 </div>
