+++ lending-admin-client/src/app/check-sheet-content-step/check-sheet-content-step.component.html
@@ -1,8 +1,8 @@
-<hot-table [data]="data" [rowHeaders]="true" [colHeaders]="true" stretchH="all">
+<hot-table [data]="model.data" [rowHeaders]="true" [colHeaders]="true" stretchH="all">
 </hot-table>
 
 <div class="btn-group">
   <button type="button" class="btn btn-secondary" previousStep>Zurück</button>
-  <button type="button" class="btn btn-secondary" nextStep (finalize)="finishStep()">Weiter</button>
+  <button type="button" class="btn btn-secondary" nextStep (finalize)="model.loadTable()">Weiter</button>
 </div>
