+++ lending-admin-client/src/app/confirmation-step/confirmation-step.component.html
@@ -1,12 +1,12 @@
-<hot-table hotId="confirmation-hot-table" class="confirmation-hot-table" [data]="items" [columns]="columns"
-           [rowHeaders]="true" [colHeaders]="columnHeaders" (beforeChange)="beforeCellChange($event)"
+<hot-table hotId="confirmation-hot-table" class="confirmation-hot-table" [data]="model.items" [columns]="model.columns"
+           [rowHeaders]="true" [colHeaders]="model.columnHeaders" (beforeChange)="beforeCellChange($event)"
            [contextMenu]="true" stretchH="all">
 </hot-table>
 
 <div class="btn-group">
   <button type="button" class="btn btn-secondary" previousStep>Zurück</button>
   <button type="button" class="btn btn-secondary" (click)="validate()">Validiere Spiele</button>
-  <button type="button" class="btn btn-secondary" [disabled]="!verificationResult.verified" nextStep (finalize)="confirmItems()">Fertigstellen</button>
+  <button type="button" class="btn btn-secondary" [disabled]="!model.verificationResult.verified" nextStep (finalize)="confirmItems()">Fertigstellen</button>
 </div>
 
 <ng-snotify></ng-snotify>
