+++ lending-admin-client/src/app/multiple-game-addition-model.service.spec.ts
@@ -1,0 +1,15 @@
+import { TestBed, inject } from '@angular/core/testing';
+
+import { MultipleGameAdditionModelService } from './multiple-game-addition-model.service';
+
+describe('MultipleGameAdditionModelService', () => {
+  beforeEach(() => {
+    TestBed.configureTestingModule({
+      providers: [MultipleGameAdditionModelService]
+    });
+  });
+
+  it('should be created', inject([MultipleGameAdditionModelService], (service: MultipleGameAdditionModelService) => {
+    expect(service).toBeTruthy();
+  }));
+});
